Make AirplaneModeBR as a separate class
to apply Single Responsibility principle
